LPS-78304 Refactor LDAPUtil to not be a Component. A public static Component is always a bad thing.
