+++ modules/apps/foundation/portal-settings/portal-settings-authentication-ldap-web/src/main/java/com/liferay/portal/settings/authentication/ldap/web/internal/util/LDAPFilterValidatorUtil.java
@@ -1,0 +1,41 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.settings.authentication.ldap.web.internal.util;
+
+import com.liferay.portal.security.ldap.validator.LDAPFilterValidator;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Shuyang Zhou
+ */
+@Component(immediate = true)
+public class LDAPFilterValidatorUtil {
+
+	public static boolean isValidFilter(String filter) {
+		return _ldapFilterValidator.isValid(filter);
+	}
+
+	@Reference(unbind = "-")
+	protected void setLDAPFilterValidator(
+		LDAPFilterValidator ldapFilterValidator) {
+
+		_ldapFilterValidator = ldapFilterValidator;
+	}
+
+	private static LDAPFilterValidator _ldapFilterValidator;
+
+}
