+++ modules/apps/foundation/portal-settings/portal-settings-authentication-ldap-web/src/main/resources/META-INF/resources/com.liferay.portal.settings.web/test_ldap_groups.jsp
@@ -53,7 +53,7 @@ if (Validator.isNull(ParamUtil.getString(request, "groupMappingGroupName")) || V
 
 String groupFilter = ParamUtil.getString(request, "importGroupSearchFilter");
 
-if (!LDAPUtil.isValidFilter(groupFilter)) {
+if (!LDAPFilterValidatorUtil.isValidFilter(groupFilter)) {
 %>
 
 	<liferay-ui:message key="please-enter-a-valid-ldap-search-filter" />
